Added new tests and made some refactoring
