+++ tests/Integration/Tests/Config/TestConfigTest.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Composer\Config\Tests\Integration\Tests\Config;
+
+final class TestConfigTest extends ConfigTest
+{
+    public function configProvider(): array
+    {
+        return [
+            [
+                \Environment\Serializer\SerializerInterface::class,
+                \Environment\Tests\Serializer\TestSerializer::class,
+            ],
+        ];
+    }
+
+    protected function getConfigName(): string
+    {
+        return 'test';
+    }
+}
