--- tests/Integration/Tests/ConfigTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Composer\Config\Tests\Integration\Tests;
-
-class ConfigTest extends PluginTestCase
-{
-
-}
